requirements.txt
ib_insync>=0.9.72
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
asyncio>=3.4.3
websockets>=10.0
requests>=2.28.0
matplotlib>=3.5.0  # for backtesting visualization